avoid basic warnings, respect clippy
